requirements.txt
nosetests
rednose
pycrypto
